requirements.txt
pandas
python-dotenv
matplotlib
mistral
numpy
scikit-learn
flask
